requirements.txt
mss~=6.1.0
pygame~=2.0.1
Pillow~=8.1.2
pywin32~=301